+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Compatibility with WordPress 5.9.
 - New archiver to store individually each email as **json** or **eml** file, on server filesystem (local hard-drive or mounted storage).
 - New archiver to forward each mail to a specified email address.
+- New archiver to store archives GMail.
+- New archiver to store archives in an Imap server.
 - New archiver to store archives in an Elasticsearch instance.
 - New archiver to store archives in Elastic Cloud.
 - New archiver to store archives in a Loki instance.